// src/generated/prisma/models/payments.ts
export type paymentsModel = runtime.Types.Result.DefaultSelection<Prisma.$paymentsPayload>
// ...
export type $paymentsPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "payments"
  objects: {
    orders: Prisma.$ordersPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    order_id: string
    payment_method: string
    mercadopago_payment_id: string | null
    mercadopago_preference_id: string | null
    payment_data: runtime.JsonValue | null
    status: string
    created_at: Date
    updated_at: Date
  }, ExtArgs["result"]["payments"]>
  composites: {}
}